# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Sea Hunter - Poker Staking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Session {
  id: string;
  shooterId: string;
  shooterUsername: string;
  gameName: string;
  buyInUsd: number;
  /** Percentage per share (e.g. 5 means each share is 5% of 75%) */
  sharePercent: number;
  /** Total number of available shares (always 75% / sharePercent) */
  totalShares: number;
  /** Number of confirmed sold shares */
  soldShares: number;
  status: SessionStatus;
  /**
   * win or loss
   * @nullable
   */
  result?: string | null;
  /**
   * Net profit/loss for the session
   * @nullable
   */
  profitUsd?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type ShareStatus = (typeof ShareStatus)[keyof typeof ShareStatus];

export const ShareStatus = {
  available: "available",
  pending: "pending",
  confirmed: "confirmed",
} as const;

export interface Share {
  id: string;
  sessionId: string;
  /** Which slot in the pie (0-indexed) */
  slotIndex: number;
  /** @nullable */
  backerId?: string | null;
  /** @nullable */
  backerUsername?: string | null;
  sharePercent: number;
  priceUsd: number;
  status: ShareStatus;
  /** @nullable */
  txId?: string | null;
  createdAt: string;
}

export type SessionDetail = Session & {
  shares: Share[];
};

export interface CreateSessionBody {
  /** @minLength 1 */
  gameName: string;
  /** @minimum 1 */
  buyInUsd: number;
  /**
   * Percentage per share of the 75% backer pool
   * @minimum 1
   * @maximum 75
   */
  sharePercent: number;
}

export type UpdateSessionBodyStatus =
  (typeof UpdateSessionBodyStatus)[keyof typeof UpdateSessionBodyStatus];

export const UpdateSessionBodyStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type UpdateSessionBodyResult =
  (typeof UpdateSessionBodyResult)[keyof typeof UpdateSessionBodyResult];

export const UpdateSessionBodyResult = {
  win: "win",
  loss: "loss",
} as const;

export interface UpdateSessionBody {
  status?: UpdateSessionBodyStatus;
  result?: UpdateSessionBodyResult;
  profitUsd?: number;
}

export interface PurchaseShareBody {
  slotIndex: number;
  /**
   * Payment transaction ID/reference
   * @minLength 1
   */
  txId: string;
}

export interface UserProfile {
  id: string;
  username: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  winStreak?: number;
  createdAt: string;
}

export interface UpdateProfileBody {
  /**
   * @minLength 2
   * @maxLength 30
   */
  username?: string;
}

export interface BackerStats {
  totalInvestedUsd: number;
  totalWonUsd: number;
  activeShares: number;
}

export interface ShooterStats {
  totalSessions: number;
  wins: number;
  losses: number;
  winStreak: number;
  bestWinStreak: number;
}

export interface UserStats {
  backerStats: BackerStats;
  shooterStats: ShooterStats;
}

export interface LeaderboardEntry {
  userId: string;
  username: string;
  /** @nullable */
  profileImageUrl: string | null;
  winStreak: number;
  totalWins: number;
  totalSessions: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListSessionsParams = {
  status?: ListSessionsStatus;
  shooterId?: string;
};

export type ListSessionsStatus =
  (typeof ListSessionsStatus)[keyof typeof ListSessionsStatus];

export const ListSessionsStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;