# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Deep Sea Hunter - Poker Staking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shares/{id}/confirm:
    patch:
      tags: [Shares]
      operationId: confirmShare
      summary: Confirm a share payment (shooter confirms)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Share confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/me:
    get:
      tags: [Users]
      operationId: getMyProfile
      summary: Get current user profile
      responses:
        '200':
          description: User profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Users]
      operationId: updateMyProfile
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/me/stats:
    get:
      tags: [Users]
      operationId: getMyStats
      summary: Get current user backer and shooter stats
      responses:
        '200':
          description: User stats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Share:
      type: object
      required: [id, sessionId, slotIndex, sharePercent, priceUsd, status, createdAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        slotIndex:
          type: integer
          description: Which slot in the pie (0-indexed)
        backerId:
          type: ['string', 'null']
        backerUsername:
          type: ['string', 'null']
        sharePercent:
          type: number
        priceUsd:
          type: number
        status:
          type: string
          enum: [available, pending, confirmed]
        txId:
          type: ['string', 'null']
        createdAt:
          type: string
          format: date-time
...
    UserProfile:
      type: object
      required: [id, username, email, profileImageUrl, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: ['string', 'null']
        profileImageUrl:
          type: ['string', 'null']
        winStreak:
          type: integer
        createdAt:
          type: string
          format: date-time
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
    UserStats:
      type: object
      required: [backerStats, shooterStats]
      properties:
        backerStats:
          $ref: '#/components/schemas/BackerStats'
        shooterStats:
          $ref: '#/components/schemas/ShooterStats'
    BackerStats:
      type: object
      required: [totalInvestedUsd, totalWonUsd, activeShares]
      properties:
        totalInvestedUsd:
          type: number
        totalWonUsd:
          type: number
        activeShares:
          type: integer
    ShooterStats:
      type: object
      required: [totalSessions, wins, losses, winStreak, bestWinStreak]
      properties:
        totalSessions:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winStreak:
          type: integer
        bestWinStreak:
          type: integer
...