# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Sea Hunter - Poker Staking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}