# lib/api-zod/src/generated/types/share.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Sea Hunter - Poker Staking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ShareStatus } from "./shareStatus";

export interface Share {
  id: string;
  sessionId: string;
  /** Which slot in the pie (0-indexed) */
  slotIndex: number;
  /** @nullable */
  backerId?: string | null;
  /** @nullable */
  backerUsername?: string | null;
  sharePercent: number;
  priceUsd: number;
  status: ShareStatus;
  /** @nullable */
  txId?: string | null;
  createdAt: Date;
}